+++ packages/mcp-server/src/auth.ts
@@ -2,8 +2,9 @@
 
 import { IncomingMessage } from 'node:http';
 import { ClientOptions } from 'sports-odds-api';
+import { McpOptions } from './options';
 
-export const parseAuthHeaders = (req: IncomingMessage, required?: boolean): Partial<ClientOptions> => {
+export const parseClientAuthHeaders = (req: IncomingMessage, required?: boolean): Partial<ClientOptions> => {
   const apiKeyHeader =
     Array.isArray(req.headers['x-api-key']) ? req.headers['x-api-key'][0] : req.headers['x-api-key'];
   const apiKeyParam =
@@ -11,4 +12,18 @@ export const parseAuthHeaders = (req: IncomingMessage, required?: boolean): Part
       req.headers['x-sports-odds-api-key-header'][0]
     : req.headers['x-sports-odds-api-key-header'];
   return { apiKeyHeader, apiKeyParam };
+};
+
+export const getStainlessApiKey = (req: IncomingMessage, mcpOptions: McpOptions): string | undefined => {
+  // Try to get the key from the x-stainless-api-key header
+  const headerKey =
+    Array.isArray(req.headers['x-stainless-api-key']) ?
+      req.headers['x-stainless-api-key'][0]
+    : req.headers['x-stainless-api-key'];
+  if (headerKey && typeof headerKey === 'string') {
+    return headerKey;
+  }
+
+  // Fall back to value set in the mcpOptions (e.g. from environment variable), if provided
+  return mcpOptions.stainlessApiKey;
 };