+++ src/resources/leagues.ts
@@ -4,6 +4,9 @@ import { APIResource } from '../core/resource';
 import { APIPromise } from '../core/api-promise';
 import { RequestOptions } from '../internal/request-options';
 
+/**
+ * Get League-related data
+ */
 export class Leagues extends APIResource {
   /**
    * Get a list of Leagues